snowflake_requirements.txt
snowflake-connector-python>=3.0.0
snowflake-sqlalchemy>=1.4.0
google-genai>=1.0.0
PyMuPDF>=1.20.0
Pillow>=9.0.0
tqdm>=4.60.0
voyageai>=0.3.0
numpy>=1.20.0
pandas>=1.3.0
sqlalchemy>=1.4.0
